# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Goat Nutrition App API
  version: 0.1.0
  description: |
    Minimal OpenAPI for key endpoints. Auth uses cookie session (Flask-Login).
servers:
  - url: http://localhost:5001
paths:
...
  /api/sheep/:
    get:
      summary: List sheep (current user)
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
    post:
      summary: Create sheep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [EarNum]
              properties:
                EarNum: { type: string }
                Breed: { type: string }
                Sex: { type: string }
                BirthDate: { type: string }
      responses:
        '201': { description: Created }
        '400': { description: Validation error }
        '409': { description: Duplicate EarNum }
  /api/sheep/{ear_num}:
    parameters:
      - in: path
        name: ear_num
        required: true
        schema: { type: string }
    get:
      summary: Get one sheep (with events)
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
    put:
      summary: Update sheep (records history for key fields)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: OK }
        '400': { description: Validation error }
        '404': { description: Not Found }
    delete:
      summary: Delete sheep
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
  /api/sheep/{ear_num}/events:
    parameters:
      - in: path
        name: ear_num
        required: true
        schema: { type: string }
    get:
      summary: List events
      responses:
        '200': { description: OK }
    post:
      summary: Add event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_date, event_type]
              properties:
                event_date: { type: string }
                event_type: { type: string }
                description: { type: string }
                medication: { type: string }
                withdrawal_days: { type: integer }
      responses:
        '201': { description: Created }
  /api/sheep/events/{event_id}:
    parameters:
      - in: path
        name: event_id
        required: true
        schema: { type: integer }
    put:
      summary: Update event
      responses:
        '200': { description: OK }
    delete:
      summary: Delete event
      responses:
        '200': { description: OK }
  /api/sheep/{ear_num}/history:
    parameters:
      - in: path
        name: ear_num
        required: true
        schema: { type: string }
    get:
      summary: List historical data
      responses:
        '200': { description: OK }
  /api/sheep/history/{record_id}:
    parameters:
      - in: path
        name: record_id
        required: true
        schema: { type: integer }
    delete:
      summary: Delete historical record
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
security:
  - cookieAuth: []